Big update
Public stash has been updated to handle beast. Also more fields has been exposed in the API

ItemMods is not finished, close though. Fixing buggs and updating the local json models to mirror path of exile's mods state. The json models will be included in later commits when itemmods is in such a state it can be used to produce new content.